* Minor changes and fine-tuning start.

+++ src/opengl_render.h
@@ -109,7 +109,7 @@ void gl_renderShaderH( const SimpleShader *shd, const mat4 *H,
    gl_renderCircle( ( x ), ( y ), ( r ), ( c ), 0 )
 void gl_renderCircle( double x, double y, double r, const glColour *c,
                       double line_width );
-void gl_renderCircleH( const mat4 *H, const glColour *c, double line_width );
+void gl_renderCircleH( const mat4 *H, const glColour *c, int line_width );
 
 /* Lines. */
 void gl_renderLine( double x1, double y1, double x2, double y2,
